requirements.txt
pytest==8.3.4
selenium==4.27.1
pyyaml==6.0.2
allure-pytest==2.13.5
